Initial platform class splitting commit.
