+++ include/singa/core/device.h
@@ -187,56 +187,6 @@ class CudaGPU : public Device {
 
 #endif  // USE_CUDA
 
-/// For querying physical devices and creating singa::Device instances.
-class Platform {
- public:
-  /// Return the number of total avaiable GPUs
-  static int GetNumGPUs();
-
-  /// Return the device IDs of available GPUs.
-  /// TODO(wangwei) return the IDs according to free memory in decending order
-  static const vector<int> GetGPUIDs();
-
-  static const std::pair<size_t, size_t> GetGPUMemSize(const int device);
-  /// Return the memory of a GPU <free, total>
-  static const vector<std::pair<size_t, size_t>> GetGPUMemSize();
-
-  /// Return a string containing all hardware info, e.g., version, memory size.
-  static const string DeviceQuery(int id, bool verbose = false);
-
-  /// Create a set of CudaGPU Device using 'num_devices' free GPUs.
-  static const vector<shared_ptr<Device> >
-  CreateCudaGPUs(const size_t num_devices, size_t init_size = 0);
-
-  /// Create a set of CudaGPU Device using given GPU IDs.
-  static const vector<shared_ptr<Device> >
-  CreateCudaGPUs(const vector<int> &devices, size_t init_size = 0);
-
-  /// Create a set of OpenclGPU Device using 'num_devices' free GPUs.
-  const vector<shared_ptr<Device>> CreateOpenclGPUs(const size_t num_devices);
-
-  /// Create a set of OpenclGPU Device using given GPU IDs.
-  const vector<shared_ptr<Device>> CreateOpenclGPUs(const vector<int>& id);
-  /// This function is implementd by Caffe (http://caffe.berkeleyvision.org/).
-  /// This function checks the availability of GPU #device_id.
-  /// It attempts to create a context on the device by calling cudaFree(0).
-  /// cudaSetDevice() alone is not sufficient to check the availability.
-  /// It lazily records device_id, however, does not initialize a
-  /// context. So it does not know if the host thread has the permission to use
-  /// the device or not.
-  ///
-  /// In a shared environment where the devices are set to EXCLUSIVE_PROCESS
-  /// or EXCLUSIVE_THREAD mode, cudaSetDevice() returns cudaSuccess
-  /// even if the device is exclusively occupied by another process or thread.
-  /// Cuda operations that initialize the context are needed to check
-  /// the permission. cudaFree(0) is one of those with no side effect,
-  /// except the context initialization.
-  static bool CheckDevice(const int device_id);
-
-// private:
-  Platform() {};  // No need to construct an instance as it has no member fields
-};
-
 }  // namespace singa
 
 #endif  // SINGA_CORE_DEVICE_H_
